requirements.txt
Flask>=2.2.0
Flask-Cors>=3.0.10
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
